# api/docs/swagger.yaml
definitions:
...
  structs.CreateSalon:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  structs.CreateTrans:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  structs.Error:
    properties:
      message:
        type: string
    type: object
...
  structs.Model:
    properties:
      created_at:
        type: string
      id:
        type: integer
      marc_id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  structs.Salon:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  structs.Salons:
    properties:
      salons:
        items:
          $ref: '#/definitions/structs.Salon'
        type: array
    type: object
  structs.StandardErrorModel:
    properties:
      error:
        $ref: '#/definitions/structs.Error'
    type: object
  structs.Transmission:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  structs.Transmissions:
    properties:
      transmissions:
        items:
          $ref: '#/definitions/structs.Transmission'
        type: array
    type: object
...
  structs.UpdateSalonReq:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  structs.UpdateTrans:
    properties:
      id:
        type: integer
      name:
        type: string
    required:
    - id
    - name
    type: object
info:
  contact: {}
paths:
...
  /v1/salon:
    post:
      consumes:
      - application/json
      description: This API for creating salon
      parameters:
      - description: CreateSalon
        in: body
        name: salon
        required: true
        schema:
          $ref: '#/definitions/structs.CreateSalon'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Salon'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: CreateSalon
      tags:
      - salon
    put:
      consumes:
      - application/json
      description: This API for updating salon for cars
      parameters:
      - description: UpdateSalon
        in: body
        name: salon
        required: true
        schema:
          $ref: '#/definitions/structs.UpdateSalonReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Salon'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      tags:
      - salon
  /v1/salon/{id}:
    delete:
      consumes:
      - application/json
      description: This API for deleting salon by id
      parameters:
      - description: Salon_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Model'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: DeleteSalon
      tags:
      - salon
    get:
      consumes:
      - application/json
      description: This API for getting salon by id
      parameters:
      - description: Salon_Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Salon'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: GetSalonById
      tags:
      - salon
  /v1/salons:
    get:
      consumes:
      - application/json
      description: This API for getting all salons
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Salons'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: GetAllSalons
      tags:
      - salon
  /v1/transmission:
    post:
      consumes:
      - application/json
      description: This API forcreating transmission
      parameters:
      - description: Transmission
        in: body
        name: transmission
        required: true
        schema:
          $ref: '#/definitions/structs.CreateTrans'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Transmission'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: CreateTransmission
      tags:
      - transmission
    put:
      consumes:
      - application/json
      description: This API for updating transmission
      parameters:
      - description: Update_Transmission
        in: body
        name: trans
        required: true
        schema:
          $ref: '#/definitions/structs.UpdateTrans'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Transmission'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: UpdateTransmission
      tags:
      - transmission
  /v1/transmission/{id}:
    delete:
      consumes:
      - application/json
      description: This API for deleting transmission
      parameters:
      - description: Transmission_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Transmission'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: DeleteTransmission
      tags:
      - transmission
    get:
      consumes:
      - application/json
      description: This API for getting transmission by id
      parameters:
      - description: Transmission_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Transmission'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: GetTransmissionById
      tags:
      - transmission
  /v1/transmissions:
    get:
      consumes:
      - application/json
      description: This API for getting all transmission
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Transmissions'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: GetAllTransmission
      tags:
      - transmission
swagger: "2.0"